Move the free routines to a better location
