+++ link-grammar/linkage/linkage.h
@@ -92,4 +92,6 @@ void partial_init_linkage(Sentence, Linkage, unsigned int N_words);
 void check_link_size(Linkage);
 void remove_empty_words(Linkage);
 void free_linkage(Linkage);
+void free_linkages(Sentence);
+
 #endif /* _LINKAGE_H */
